Split MATLAB generator into two variants
